snippets/python/pyproject.toml
[project]
name = "python_docs_snippets"
version = "0.1.0"
description = "Python code snippets for documentation"
requires-python = ">=3.11"
dependencies = [
    "restate_sdk[serde]==0.9.1",
    "hypercorn",
    "pydantic",
    "requests",
    "types-requests",
    "fastapi"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
